Split the analysis code from the jenkins query code.
